+++ xfields/ibs/__init__.py
@@ -6,7 +6,7 @@
 from ._analytical import BjorkenMtingwaIBS, IBSGrowthRates, NagaitsevIBS
 from ._api import configure_intrabeam_scattering, get_intrabeam_scattering_growth_rates
 from ._kicks import IBSAnalyticalKick, IBSKineticKick
-from ._equilibrium_emittance import compute_emittance_evolution
+from ._equilibrium import compute_emittance_evolution
 
 __all__ = [
     "BjorkenMtingwaIBS",
